django_project/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
crispy-bootstrap4==2022.1
Django==4.2.4
django-crispy-forms==2.0
idna==3.4
Pillow==10.0.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4

# for deployment on Ubuntu
# gunicorn==21.2.0
# psycopg2-binary==2.9.9
